[mv + rename] erp 관련 파일들 pages 폴더 하위로 이관 → chatbot 관련 파일명 rename
